More pie_BoxFill* -Wfloat-conversion fixes

+++ lib/ivis_opengl/piedraw.cpp
@@ -844,15 +844,15 @@ static ShadowCache shadowCache;
 
 static void pie_DrawShadows(uint64_t currentGameFrame)
 {
-	const float width = pie_GetVideoBufferWidth();
-	const float height = pie_GetVideoBufferHeight();
+	const int width = pie_GetVideoBufferWidth();
+	const int height = pie_GetVideoBufferHeight();
 	shadowCache.setCurrentFrame(currentGameFrame);
 
 	pie_ShadowDrawLoop(shadowCache);
 
 	PIELIGHT grey;
 	grey.byte = { 0, 0, 0, 128 };
-	pie_BoxFill_alpha(0.f, 0.f, width, height, grey);
+	pie_BoxFill_alpha(0, 0, width, height, grey);
 
 	scshapes.resize(0);
 	shadowCache.removeUnused();
